chore: bump metakit to 1.8.0-rc.0 (#157)
* chore: bump metakit to 1.8.0-rc.0

FloatValue is now Ratio-backed (exact-rational numerics): the
feeCalculated test extraction uses Ratio.toBigInt via RatioOps.


* chore: pin metakit 1.8.0-rc.2


* chore: pin metakit 1.8.0-rc.1

Tag was reused after the doc-jar release fix; artifact verified live
on Maven Central.


* fix(ci): map tessellation v4.0.0 to rc.10 release assets in e2e

The v4.0.0 git tag has no GitHub release with jar assets; rc.10 has
the full set and is version-stamp-only away from final. Remove once
Constellation publishes v4.0.0 assets.


* fix(shared-data): adapt float handling to exact-Ratio FloatValue

metakit 1.8 unified the numeric model: FloatValue wraps an exact Ratio,
whose toString prints "n/d". Adjustments:

- CommonRules.estimateValueSize: size FloatValue by its decimal rendering
  (Ratio.toBigDecimal), matching the serialized wire form, instead of the
  rational "n/d" string;
- ScriptStateMachineIntegrationSuite: extract BigInt via Ratio.toBigInt
  (the old .toLong round-trip truncated through Double);
- RiverdaleEconomyStateMachineSuite: compare extracted floats against
  exact Ratio values (Ratio.fromBigDecimal / Ratio.Zero), never Double
  literals.

No gas-sensitive expectations needed re-baselining for the corrected
charge-once metering: ottochain suites assert gas bounds and behavior,
not exact charges (the only literal, gasLimit == 50L, echoes a config
value), and the full suite is green against the new metering unchanged.

(cherry picked from commit a6ae302ef553f8a8e8e0dc9552a716cc0f4579ae)

* fix(e2e): drop nulls before signing to match metakit 1.8 canonical

Every e2e scenario failed at step 1 with HTTP 400 from the DL1 /data
endpoint after the metakit 1.7.0-rc.9 -> 1.8.0-rc.1 bump.

Root cause: metakit 1.8 (PR #27) restored drop-nulls-before-canonicalize
in JsonBinaryCodec.serialize, so the node's serializeUpdate — the bytes
DataApplicationRoutes hashes to verify signature proofs — no longer
contains null object fields. The e2e client signs over raw RFC 8785
canonical JSON that DOES contain explicit nulls ("parentFiberId": null
from createFiber.ts, "initialState": null from createScript.ts, and
"metadata"/"effect": null inside fixture definitions), a shape that was
deliberately aligned to metakit 1.7's no-drop canonical form. The hash
mismatch made every create/createScript fail signature verification
(InvalidSignature -> BadRequest). Depth guard (fixtures nest ~26 < 64),
size limits, and codec strictness were ruled out: fixture payloads
decode and pass L1 validation.

Fix: apply the SDK's dropNulls to the message in sendDataTransaction
before batchSign and submit the null-free value — the single chokepoint
all scenario traffic goes through. Update the stale createScript.ts
comment documenting the obsolete 1.7 contract.

E2eSignedPayloadCompatSuite pins the contract at the byte level using
the actual voting and counter-script fixtures: client bytes with nulls
diverge from node serializeUpdate bytes, drop-nulls client bytes match
exactly, and the null-free wire value decodes identically.


* fix(e2e): inline dropNulls (vendored sdk lacks the export)


---------

+++ modules/shared-data/src/main/scala/xyz/kd5ujc/shared_data/lifecycle/validate/rules/CommonRules.scala
@@ -8,6 +8,7 @@ import cats.syntax.all._
 
 import io.constellationnetwork.currency.dataApplication.DataApplicationValidationError
 import io.constellationnetwork.metagraph_sdk.json_logic._
+import io.constellationnetwork.metagraph_sdk.numerics.RatioOps.implicits._
 
 import xyz.kd5ujc.schema.OnChain
 import xyz.kd5ujc.shared_data.lifecycle.validate.{Limits, ValidationResult}
@@ -129,10 +130,13 @@ object CommonRules {
     if (depth > 100) return 100L // Prevent infinite recursion
 
     value match {
-      case NullValue     => 4L
-      case BoolValue(_)  => 5L
-      case IntValue(v)   => 8L + v.toString.length
-      case FloatValue(v) => 8L + v.toString.length
+      case NullValue    => 4L
+      case BoolValue(_) => 5L
+      case IntValue(v)  => 8L + v.toString.length
+      // FloatValue wraps an exact Ratio whose toString prints "n/d" (e.g. "1/3");
+      // size it by the decimal rendering used on the wire (Json.fromBigDecimal)
+      // so estimates stay consistent with the serialized payload.
+      case FloatValue(v) => 8L + v.toBigDecimal.toString.length
       case StrValue(s)   => 4L + s.length.toLong * 2L // UTF-16 encoding estimate
       case ArrayValue(items) =>
         8L + items.map(estimateValueSize(_, depth + 1)).sum
